+++ common/src/main/java/tachyon/network/protocol/EncodedMessage.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.worker.netty.protocol;
+package tachyon.network.protocol;
 
 import io.netty.buffer.ByteBuf;
