+++ common/src/main/java/tachyon/network/protocol/RPCMessageDecoder.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.worker.netty.protocol;
+package tachyon.network.protocol;
 
 import java.util.List;
