+++ common/src/main/java/tachyon/network/protocol/RPCResponse.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.worker.netty.protocol;
+package tachyon.network.protocol;
 
 public abstract class RPCResponse extends RPCMessage {
 }
