+++ common/src/main/java/tachyon/network/protocol/buffer/DataFileChannel.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.worker.netty.protocol.buffer;
+package tachyon.network.protocol.buffer;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
