+++ integration-tests/src/test/java/tachyon/worker/netty/protocol/MessageSavingHandler.java
@@ -23,6 +23,7 @@ import org.junit.Assert;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
+import tachyon.network.protocol.RPCMessage;
 
 /**
  * This is a simple handler for the channel pipelines. This simply saves the message it receives.
