+++ servers/src/main/java/tachyon/worker/package-info.java
@@ -52,8 +52,8 @@
  * {@link tachyon.worker.DataServerMessage#createBlockResponseMessage(boolean, long,
  * java.nio.ByteBuffer)}.
  * Side note, the netty implementation does not use this class, but has defined two classes for
- * the read and write case: {@link tachyon.worker.netty.protocol.RPCBlockRequest},
- * {@link tachyon.worker.netty.protocol.RPCBlockResponse}; theses classes are network compatible.
+ * the read and write case: {@link tachyon.network.protocol.RPCBlockRequest},
+ * {@link tachyon.network.protocol.RPCBlockResponse}; theses classes are network compatible.
  */
 package tachyon.worker;
