+++ servers/src/test/java/tachyon/worker/netty/protocol/RPCBlockRequestTest.java
@@ -21,6 +21,8 @@ import org.junit.Test;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
+import tachyon.network.protocol.RPCBlockRequest;
+import tachyon.network.protocol.RPCMessage;
 
 public class RPCBlockRequestTest {
   private static final long BLOCK_ID = 11;
