+++ servers/src/test/java/tachyon/worker/netty/protocol/buffer/DataByteBufferTest.java
@@ -25,6 +25,7 @@ import io.netty.buffer.ByteBuf;
 import io.netty.channel.FileRegion;
 
 import tachyon.TestUtils;
+import tachyon.network.protocol.buffer.DataByteBuffer;
 
 public class DataByteBufferTest {
   private static final int LENGTH = 5;
